changed the code file struture
