+++ src/components/cqlLibraryRoutes/CqlLibraryRoutes.test.tsx
@@ -3,9 +3,9 @@ import "@testing-library/jest-dom";
 
 import * as React from "react";
 import { render, screen } from "@testing-library/react";
-import ApiContextProvider from "../api/ServiceContext";
-import CqlLibraryRoutes from "./cqlLibraryRoutes/CqlLibraryRoutes";
-import { CqlLibraryServiceApi } from "../api/useCqlLibraryServiceApi";
+import CqlLibraryRoutes from "./CqlLibraryRoutes";
+import { ServiceConfig } from "../../api/ServiceContext";
+import { CqlLibraryServiceApi } from "../../api/useCqlLibraryServiceApi";
 
 const cqlLibrary = [
   {
@@ -18,7 +18,19 @@ const cqlLibrary = [
   },
 ];
 
-jest.mock("../hooks/useOktaTokens", () => () => ({
+const serviceConfig: ServiceConfig = {
+  measureService: {
+    baseUrl: "example-service-url",
+  },
+  elmTranslationService: {
+    baseUrl: "test-elm-service",
+  },
+  cqlLibraryService: {
+    baseUrl: "example-service-url",
+  },
+};
+
+jest.mock("../../hooks/useOktaTokens", () => () => ({
   getAccessToken: () => "test.jwt",
 }));
 
@@ -26,11 +38,11 @@ const mockCqlLibraryServiceApi = {
   fetchCqlLibraries: jest.fn().mockResolvedValue(cqlLibrary),
 } as unknown as CqlLibraryServiceApi;
 
-jest.mock("../api/useCqlLibraryServiceApi", () =>
+jest.mock("../../api/useCqlLibraryServiceApi", () =>
   jest.fn(() => mockCqlLibraryServiceApi)
 );
 
-describe("Displaying CQL Library Routes component", () => {
+describe("CqlLibraryRoutes", () => {
   test("shows the children when the checkbox is checked", async () => {
     render(
       <div id="main">
